# openapi.json
{
  "components": {
    "schemas": {
...
      "CapabilityListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntityRefItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CapabilityListResponse",
        "type": "object"
      },
...
      "CreateCapabilityRequest": {
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "title": "Attributes",
            "type": "object"
          },
          "capability_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capability Type"
          },
          "entity_type": {
            "const": "capability",
            "default": "capability",
            "title": "Entity Type",
            "type": "string"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateCapabilityRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "opaque",
        "description": "Paste an API token.\n\n- **Local development:** run `make dev-token` \u2014 it seeds a tenant + actor + role grants and prints a fresh token. No env-var setup required if the docker-compose stack is up.\n- **Production / break-glass:** mint with `python scripts/mint_token.py --tenant-id <uuid> --actor-id <uuid> --roles <role>`. Requires the tenant and actor to already exist.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "title": "registry",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "description": "Paginated list of capabilities visible to the caller's tenant.\n\nPagination is keyset-based: the response carries ``next_cursor`` (or null\nwhen no further pages exist). Pass ``cursor=<value>`` on the next request\nto retrieve the following page.\n\nThe legacy ``?page=N`` offset parameter is no longer accepted. Clients\nthat send it receive a 422 with code ``page_param_deprecated``.",
        "operationId": "list_capabilities_v1_capabilities_get",
        "parameters": [
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle"
            }
          },
          {
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "description": "Opaque cursor returned by the previous page. Omit to start from the first page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor returned by the previous page. Omit to start from the first page.",
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "ISO-8601 UTC datetime for time-travel",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 UTC datetime for time-travel",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Capabilities",
        "tags": [
          "retrieval"
        ]
      },
      "post": {
        "description": "Create a new capability.\n\nHonours ``X-Idempotency-Key`` (optional). Resend with the same key\n+ same body \u2192 returns the original response. Same key + different\nbody \u2192 409 with ``code: \"idempotency_key_conflict\"``.",
        "operationId": "create_capability_v1_capabilities_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCapabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Capability",
        "tags": [
          "capabilities"
        ]
      }
    },
...
    "/v1/capabilities/{entity_id}": {
      "delete": {
        "description": "Soft-delete idempotency:\n- Row exists (active or already-invalidated) \u2192 204 No Content.\n- Row never existed \u2192 404 Not Found (service raises NotFoundError).",
        "operationId": "delete_capability_v1_capabilities__entity_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Capability",
        "tags": [
          "capabilities"
        ]
      },
      "get": {
        "description": "Return the full capability record.\n\nThe path segment accepts either a UUID or a slug-form name \u2014 they\nresolve to the same record. Slugs are case-insensitive against the\nstored `name` column.\n\nOptional `?as_of=` activates bi-temporal time-travel.\n\nOptional `?include=` adds bounded sub-resources to the response. Use\nthis to collapse \"fetch capability + components + facts + external\nIDs\" from four round-trips into one.\n\nOptional `?view=audit` returns the full bitemporal + tenant-id +\nsupersession audit shape. Default `view=default` omits those fields.",
        "operationId": "get_capability_v1_capabilities__entity_id__get",
        "parameters": [
          {
            "description": "Capability UUID or slug-form name (e.g. 'salt-design-system')",
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "description": "Capability UUID or slug-form name (e.g. 'salt-design-system')",
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "description": "ISO-8601 UTC datetime for bi-temporal time-travel",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 UTC datetime for bi-temporal time-travel",
              "title": "As Of"
            }
          },
          {
            "description": "Comma-separated list of sub-resources to expand. Known values: components, depends_on, external_ids, interface. Each expansion is capped at 200 items \u2014 `truncated: true` + a `next` URL signal overflow.",
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of sub-resources to expand. Known values: components, depends_on, external_ids, interface. Each expansion is capped at 200 items \u2014 `truncated: true` + a `next` URL signal overflow.",
              "title": "Include"
            }
          },
          {
            "description": "Response shape. `default` (UI-flavoured) is the standard minimal shape every endpoint returns. `audit` adds bitemporal columns (valid_from / valid_to / ingested_at / invalidated_at), tenant_id, and supersession metadata for audit / compliance consumers.",
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "default",
              "description": "Response shape. `default` (UI-flavoured) is the standard minimal shape every endpoint returns. `audit` adds bitemporal columns (valid_from / valid_to / ingested_at / invalidated_at), tenant_id, and supersession metadata for audit / compliance consumers.",
              "title": "View",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated list of fact categories to include in the `facts` field (e.g. `release_note,overview`). Default: no filter \u2014 every category is returned. Use this to narrow a fat detail response, e.g. fetch Salt + just its release notes in one call.",
            "in": "query",
            "name": "facts_categories",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of fact categories to include in the `facts` field (e.g. `release_note,overview`). Default: no filter \u2014 every category is returned. Use this to narrow a fat detail response, e.g. fetch Salt + just its release notes in one call.",
              "title": "Facts Categories"
            }
          },
          {
            "description": "Cap the number of facts returned (applied after the category filter). Default: no cap. Useful when a capability has hundreds of facts and the UI only renders the top N.",
            "in": "query",
            "name": "facts_limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 500,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cap the number of facts returned (applied after the category filter). Default: no cap. Useful when a capability has hundreds of facts and the UI only renders the top N.",
              "title": "Facts Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Capability",
        "tags": [
          "capabilities"
        ]
      },
      "patch": {
        "description": "Update mutable attributes on a capability.\n\nHonours the ``If-Match`` request header (advisory): if present and stale,\nreturns 412 Precondition Failed; if absent, logs a warning and accepts the\nwrite.  ETag is computed from the entity row before the write so a stale\nprecondition fails fast without executing the mutation.\n\nWhen there is no detail GET that a client can use to acquire the ETag,\nthe client may compute it from the list response or from a prior PATCH\nresponse body \u2014 but that is uncommon.  The recommended flow is:\nGET /v1/capabilities/{id} \u2192 ETag header \u2192 PATCH with If-Match.",
        "operationId": "patch_capability_v1_capabilities__entity_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Capability",
        "tags": [
          "capabilities"
        ]
      }
    },
...
    "/v1/capabilities/{provider_cap_id}/adoptions": {
      "get": {
        "description": "Return the calling tenant's active adoption for this capability,\nif any.\n\nScoped to the caller's tenant \u2014 listing other tenants' adoptions for\nthe same capability is not supported through this endpoint (use the\nprojection endpoints for the provider-side view).\n\nPass ``?view=audit`` to include bitemporal columns in the response.\n\nPagination: ``next_cursor`` is always ``None`` \u2014 adoptions per capability\nper tenant are bounded (at most one active row), so keyset pagination is\nnot wired. The envelope exists for client shape consistency.",
        "operationId": "list_adoptions_v1_capabilities__provider_cap_id__adoptions_get",
        "parameters": [
          {
            "description": "Provider capability UUID or slug",
            "in": "path",
            "name": "provider_cap_id",
            "required": true,
            "schema": {
              "description": "Provider capability UUID or slug",
              "title": "Provider Cap Id",
              "type": "string"
            }
          },
          {
            "description": "Response shape. ``default`` is the standard UI-flavoured shape. ``audit`` adds bitemporal columns (valid_from / valid_to / ingested_at / invalidated_at) for audit / compliance consumers.",
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "default",
              "description": "Response shape. ``default`` is the standard UI-flavoured shape. ``audit`` adds bitemporal columns (valid_from / valid_to / ingested_at / invalidated_at) for audit / compliance consumers.",
              "title": "View",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdoptionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List active adoptions for a capability",
        "tags": [
          "adoptions"
        ]
      },
      "post": {
        "description": "Record an adoption event + provides_to edge.\n\nThe path segment accepts a UUID or a slug-form name. The consumer\ntenant is ``ctx.tenant_id``. Returns ``201`` with the newly-created\nadoption row. ``409`` if an active adoption already exists for the\n(consumer, capability) pair (uniqueness constraint).\n\nPass ``?view=audit`` to include bitemporal columns in the response.\nHonours ``X-Idempotency-Key``: same key + same body replays the\noriginal response; same key + different body returns 409.",
        "operationId": "adopt_capability_v1_capabilities__provider_cap_id__adoptions_post",
        "parameters": [
          {
            "description": "Provider capability UUID or slug",
            "in": "path",
            "name": "provider_cap_id",
            "required": true,
            "schema": {
              "description": "Provider capability UUID or slug",
              "title": "Provider Cap Id",
              "type": "string"
            }
          },
          {
            "description": "Response shape. ``default`` is the standard UI-flavoured shape. ``audit`` adds bitemporal columns (valid_from / valid_to / ingested_at / invalidated_at) for audit / compliance consumers.",
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "default",
              "description": "Response shape. ``default`` is the standard UI-flavoured shape. ``audit`` adds bitemporal columns (valid_from / valid_to / ingested_at / invalidated_at) for audit / compliance consumers.",
              "title": "View",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdoptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdoptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adopt a provider capability (cross-tenant)",
        "tags": [
          "adoptions"
        ]
      }
    },
    "/v1/capabilities/{provider_cap_id}/adoptions/{adoption_id}": {
      "delete": {
        "description": "Soft-delete by setting t_invalidated_at. The provides_to edge is\nretained so historical bi-temporal traversal still surfaces the\nrelationship.\n\nIdempotent: calling on an already-invalidated adoption is a no-op\n(returns 204).",
        "operationId": "_unadopt_capability_v1_capabilities__provider_cap_id__adoptions__adoption_id__delete",
        "parameters": [
          {
            "description": "Provider capability UUID or slug",
            "in": "path",
            "name": "provider_cap_id",
            "required": true,
            "schema": {
              "description": "Provider capability UUID or slug",
              "title": "Provider Cap Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "adoption_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Adoption Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Soft-delete (unadopt) an adoption",
        "tags": [
          "adoptions"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}